refactor "Auth 관련 상수 도입"
