+++ src/main/java/challenging/application/auth/configuration/SecurityConfiguration.java
@@ -7,7 +7,7 @@ import challenging.application.auth.jwt.JWTUtils;
 import challenging.application.auth.oauth.OAuth2SuccessHandler;
 import challenging.application.auth.oauth.OAuth2UserServiceImpl;
 import challenging.application.auth.repository.RefreshTokenRepository;
-import challenging.application.auth.servletUtils.jwtUtils.JWTResponseUtils;
+import challenging.application.auth.utils.servletUtils.jwtUtils.JWTResponseUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
