+++ src/main/java/challenging/application/auth/oauth/OAuth2SuccessHandler.java
@@ -3,7 +3,7 @@ package challenging.application.auth.oauth;
 import challenging.application.auth.domain.RefreshToken;
 import challenging.application.auth.jwt.JWTUtils;
 import challenging.application.auth.repository.RefreshTokenRepository;
-import challenging.application.auth.servletUtils.cookie.CookieUtils;
+import challenging.application.auth.utils.servletUtils.cookie.CookieUtils;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -21,6 +21,8 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.Iterator;
 
+import static challenging.application.auth.utils.AuthConstant.*;
+
 @Component
 @RequiredArgsConstructor
 @Slf4j
@@ -37,9 +39,9 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 
         String role = getRole(authentication);
 
-        String accessToken = jwtUtil.createJwt("access", email, role, 60*60*60*60L);
+        String accessToken = jwtUtil.generateAccessToken(email, role);
 
-        String refreshToken = jwtUtil.createJwt("refresh", email, role, 1L);
+        String refreshToken = jwtUtil.generateRefreshToken(email, role);
 
         addRefreshEntity(email, refreshToken, 1L);
 
@@ -50,8 +52,8 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
     }
 
     private void setInformationInResponse(HttpServletResponse response, String accessToken, String refreshToken) throws IOException {
-        Cookie access = CookieUtils.createCookie("access", accessToken);
-        Cookie refresh = CookieUtils.createCookie("refresh", refreshToken);
+        Cookie access = CookieUtils.createCookie(ACCESS_TOKEN, accessToken);
+        Cookie refresh = CookieUtils.createCookie(REFRESH_TOKEN, refreshToken);
 
         response.addCookie(access);
         response.addCookie(refresh);
